refactor: psalm ledger to zero — baseline deleted, both analyzers run bare
Burns the remaining 974 psalm findings (of the original 7,754 across 253
files) to zero and deletes psalm-baseline.xml. psalm.xml no longer carries
an errorBaseline attribute, and AnalyzerBaselineRatchetTest now asserts
both baselines stay deleted, mirroring the phpstan side from R6.

Real bugs surfaced and fixed by the burn (all latent, all verified):
- Fleet form lat/lng wrote nonexistent machine columns — user-entered
  coordinates were silently discarded; now persist to last_location_*.
- MachineAssignmentController: both endpoints called a mineAreas()
  relation (and pivot table) that never existed — every request fataled;
  rewritten against the real machine_area_assignments ledger.
- GeofenceCrossingDetectionJob checked a phantom exited_at attribute, so
  machine re-entries were never recorded and exits re-fired every poll;
  GeofenceExitDetected broadcast the same phantom (always null).
- MaintenanceRecordController CSV export read scheduled_at (real column:
  scheduled_date) — null->format() fatal whenever the branch ran.
- RoutePlanning serialized geofence_type from a nonexistent attribute
  (real column: type) — zone typing reached the map as null.
- User::hasRole(Role) cast the model's object properties to an array for
  a whereIn — could never match; now compares roles.id.
- GeofenceDetail used DB::raw without importing the facade (class-not-
  found at runtime); AnomalyDetector location checks read nonexistent
  latitude/longitude so the whole anomaly branch was dead.
- MaintenanceRecordFactory produced inverted completed_at date ranges,
  the source of the suite's intermittent one-test flake.

Recurring mechanical work: DB-truth docblocks for nullable columns
(verified against the live schema), decimal-cast numeric-string math made
explicit, groupBy/selectRaw aggregate aliases read via getAttribute(),
stale duplicate inner model docblocks deleted (FuelTransaction,
MaintenanceRecord), Jetstream Team::users() restated with concrete
generics, closure params typed against their runtime receivers, and
config reads laundered through ApiPayload::str()/strings() where psalm's
plugin folds analysis-env values into dead-code verdicts.

psalm: No errors found (bare). phpstan: no errors (bare, since R6).
Suite: 639 passing.

+++ app/Services/AI/CostAnalyzerAgent.php
@@ -20,12 +20,12 @@ class CostAnalyzerAgent
         $insights = [];
 
         // Analyze fuel costs
-        $fuelCosts = FuelTransaction::where('team_id', $team->id)
+        $fuelCosts = (float) FuelTransaction::where('team_id', $team->id)
             ->whereDate('transaction_date', '>=', now()->subDays(30))
             ->sum('total_cost');
 
         // Analyze maintenance costs
-        $maintenanceCosts = MaintenanceRecord::where('team_id', $team->id)
+        $maintenanceCosts = (float) MaintenanceRecord::where('team_id', $team->id)
             ->whereDate('completed_at', '>=', now()->subDays(30))
             ->sum('cost');
 
@@ -39,7 +39,7 @@ class CostAnalyzerAgent
                 'title' => 'High Operational Costs',
                 'description' => 'Daily operational costs averaging R'.number_format($avgDailyCost, 2).'. Review efficiency measures.',
                 'confidence_score' => 0.85,
-                'estimated_savings' => $avgDailyCost * 0.15 * 30,
+                'estimated_savings' => $avgDailyCost * 0.15 * 30.0,
                 'data' => [
                     'total_monthly_cost' => round($totalCosts, 2),
                     'fuel_costs' => round($fuelCosts, 2),
